Chore: Move peerDeps and versionCompare into lib/utils
Both are utility modules that belong alongside the other utils, not
sitting directly in lib/ next to the classes.
