+++ bin/deps-gen.js
@@ -5,7 +5,7 @@ import CliCommand from '../lib/CliCommand.js'
 import exec from '../lib/utils/exec.js'
 import getModuleDirs from '../lib/utils/getModuleDirs.js'
 import isModule from '../lib/utils/isModule.js'
-import { isAdaptModule, deriveExpectedPeerDeps, deriveExpectedDeps, findOutdatedVersions } from '../lib/peerDeps.js'
+import { isAdaptModule, deriveExpectedPeerDeps, deriveExpectedDeps, findOutdatedVersions } from '../lib/utils/peerDeps.js'
 
 const CORE_PKG = 'adapt-authoring-core'
