+++ lib/utils/getReleases.js
@@ -1,6 +1,6 @@
 import githubRequest from './githubRequest.js'
 import semver from 'semver'
-import { normalizeBump } from '../versionCompare.js'
+import { normalizeBump } from './versionCompare.js'
 
 export default async function getReleases (options) {
   const releases = (await githubRequest('releases?per_page=10'))
