+++ lib/utils/peerDeps.js
@@ -1,6 +1,6 @@
 import { readFileSync, existsSync, statSync } from 'node:fs'
 import { join } from 'node:path'
-import collectJsFiles from './utils/collectJsFiles.js'
+import collectJsFiles from './collectJsFiles.js'
 
 export const PREFIX = 'adapt-authoring-'
